+++ browser-service/src/main/java/com/platon/browser/config/bean/Common.java
@@ -3,11 +3,13 @@ package com.platon.browser.config.bean;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class Common {
-    private int ExpectedMinutes;
-    private int NodeBlockTimeWindow;
-    private int PerRoundBlocks;
-    private int ValidatorCount;
-    private int AdditionalCycleTime;
+    private BigInteger ExpectedMinutes;
+    private BigInteger NodeBlockTimeWindow;
+    private BigInteger PerRoundBlocks;
+    private BigInteger ValidatorCount;
+    private BigInteger AdditionalCycleTime;
 }
