+++ browser-service/src/main/java/com/platon/browser/config/bean/Reward.java
@@ -3,8 +3,10 @@ package com.platon.browser.config.bean;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class Reward {
-    private int NewBlockRate;
-    private int PlatONFoundationYear;
+    private BigDecimal NewBlockRate;
+    private BigDecimal PlatONFoundationYear;
 }
